run部分微调；暴露出add_user api
